\name{A.gro}
\docType{data}
\alias{A.gro}
\title{Transcription rate data from yeast}
\description{
  This data set gives the transcription rates of 200 genes from yeast measured at two different
  conditions (glucose vs glucose starvation).
}
\usage{multiyeast}
\format{A matrix containing 6 rows and 200 variables.}
\source{Gene Expression Omnimus: GSE11521}
\references{

[1] Pelechano, V. and Pérez‐Ortín, J.E. (2010), There is a steady‐state transcriptome in exponentially growing yeast cells. Yeast, 27: 413-422. https://doi.org/10.1002/yea.1768

[2] García-Martínez, J. and Aranda, A. and Pérez-Ortín, J. E. (2004). Genomic Run-On Evaluates Transcription Rates for All Yeast Genes and Identifies Gene Regulatory Mechanisms, Molecular Cell, 15. https://doi.org/10.1016/j.molcel.2004.06.004

[3] Pelechano, V., Chávez, S., & Pérez-Ortín, J. E. (2010). A complete set of nascent transcription rates for yeast genes. PloS one, 5(11), e15442. https://doi.org/10.1371/journal.pone.0015442
}
\keyword{datasets}
