\name{A.rna}
\docType{data}
\alias{A.rna}
\title{Gene expression data from yeast}
\description{
  This data set gives the expression of 200 genes from yeast measured at two different
  conditions (glucose vs glucose starvation).
}
\usage{multiyeast}
\format{A matrix containing 6 rows and 200 variables.}
\source{Gene Expression Omnimus: GSE11521}
\references{
@article{JE1,
author = {Pelechano, Vicent and P{\'e}rez-Ort{\'i}n, Jos{\'e} E.},
title = {There is a steady-state transcriptome in exponentially growing yeast cells},
journal = {Yeast},
volume = {27},
number = {7},
pages = {413-422},
keywords = {transcription rate, steady state, transcriptome, Saccharomyces cerevisiae, mRNA stability},
doi = {10.1002/yea.1768},
url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/yea.1768},
eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/yea.1768},
year = {2010}
}

@article{JE2,
	Annote = {doi: 10.1016/j.molcel.2004.06.004},
	Author = {Garc{\'i}a-Mart{\'i}nez, Jos{\'e} and Aranda, Agust{\'i}n and P{\'e}rez-Ort{\'i}n, Jos{\'e}E},
	Booktitle = {Molecular Cell},
	Date = {2004/07/23},
	Date-Added = {2019-01-29 12:15:03 +0000},
	Date-Modified = {2019-01-29 12:15:03 +0000},
	Doi = {10.1016/j.molcel.2004.06.004},
	Isbn = {1097-2765},
	Journal = {Molecular Cell},
	M3 = {doi: 10.1016/j.molcel.2004.06.004},
	Month = {2019/01/29},
	Number = {2},
	Pages = {303--313},
	Publisher = {Elsevier},
	Title = {Genomic Run-On Evaluates Transcription Rates for All Yeast Genes and Identifies Gene Regulatory Mechanisms},
	Ty = {JOUR},
	Url = {https://doi.org/10.1016/j.molcel.2004.06.004},
	Volume = {15},
	Year = {2004},
	Year1 = {2004},
	Bdsk-Url-1 = {https://doi.org/10.1016/j.molcel.2004.06.004},
	Bdsk-Url-2 = {http://dx.doi.org/10.1016/j.molcel.2004.06.004}}

@article{JE3,
	An = {21103382},
	Author = {Pelechano, Vicent and Ch{\'a}vez, Sebasti{\'a}n and P{\'e}rez-Ort{\'i}n, Jos{\'e} E.},
	Date = {2010/11/16},
	Date-Added = {2019-01-29 12:15:29 +0000},
	Date-Modified = {2019-01-29 12:15:29 +0000},
	Db = {PubMed},
	Doi = {10.1371/journal.pone.0015442},
	Isbn = {1932-6203},
	J1 = {PLoS One},
	Journal = {PloS one},
	L2 = {https://www.ncbi.nlm.nih.gov/pmc/PMC2982843/},
	Month = {11},
	Number = {11},
	Pages = {e15442; e15442--e15442},
	Publisher = {Public Library of Science},
	Title = {A complete set of nascent transcription rates for yeast genes},
	Ty = {JOUR},
	U1 = {21103382{$[$}pmid{$]$}},
	U2 = {PMC2982843{$[$}pmcid{$]$}},
	Url = {https://www.ncbi.nlm.nih.gov/pubmed/21103382},
	Volume = {5},
	Year = {2010},
	Bdsk-Url-1 = {https://www.ncbi.nlm.nih.gov/pubmed/21103382},
	Bdsk-Url-2 = {http://dx.doi.org/10.1371/journal.pone.0015442}}
}
\keyword{datasets}
