\name{B.rna}
\docType{data}
\alias{B.rna}
\title{Gene expression data from yeast}
\description{
  This data set gives the expression of 200 genes from yeast measured at two different
  conditions (glucose vs glucose starvation).
}
\usage{multiyeast}
\format{A matrix containing 4 rows and 200 variables.}
\source{Gene Expression Omnimus: GSE56622}
\references{
@article{RIBO,
	An = {25119046},
	Author = {Zid, Brian M and O'Shea, Erin K},
	Date = {2014/10/02},
	Date-Added = {2019-05-15 08:08:55 +0000},
	Date-Modified = {2019-05-15 08:08:55 +0000},
	Db = {PubMed},
	Doi = {10.1038/nature13578},
	Et = {2014/08/03},
	Isbn = {1476-4687; 0028-0836},
	J1 = {Nature},
	Journal = {Nature},
	L2 = {https://www.ncbi.nlm.nih.gov/pmc/PMC4184922/},
	La = {eng},
	Month = {10},
	Number = {7520},
	Pages = {117--121},
	Title = {Promoter sequences direct cytoplasmic localization and translation of mRNAs during starvation in yeast},
	Ty = {JOUR},
	U1 = {25119046{$[$}pmid{$]$}},
	U2 = {PMC4184922{$[$}pmcid{$]$}},
	U4 = {nature13578{$[$}PII{$]$}},
	Url = {https://www.ncbi.nlm.nih.gov/pubmed/25119046},
	Volume = {514},
	Year = {2014},
	Bdsk-Url-1 = {https://www.ncbi.nlm.nih.gov/pubmed/25119046},
	Bdsk-Url-2 = {http://dx.doi.org/10.1038/nature13578}}
}
\keyword{datasets}

