% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiBaC.R
\name{batchCorrection}
\alias{batchCorrection}
\title{batchCorrection}
\usage{
batchCorrection(multiBatchDesign, cond.factor, Fac = NULL)
}
\arguments{
\item{multiBatchDesign}{A list containing the original and predicted omic for each batch. All omics must be present in every batch. Output object of genMissingOmics function}

\item{cond.factor}{A list with one slot for each batch containing the common experimental setting.}

\item{Fac}{Vector with numbers of components to extract in each submodel, by order:
\enumerate{
    \item Model a (time)
    \item Model b (second factor)
    \item Model c (third factor)
    \item Model ab (interaction) to not consider interations set to 0
    \item Model ac (interaction) to not consider interations set to 0
    \item Model bc (interaction) to not consider interations set to 0
    \item Model abc (interaction) to not consider interations set to 0
    \item Number components of residues
}
If NULL (default) the function optimize each value.}
}
\value{
A list with tow slots. The first one, "correctedOmics", has the same structure than multiBatchDesign input object, but containing corrected information.
The second slot contains a list with the batch effect estimation of each omic.
}
\description{
This function performs the ARSyN correction [1] for each omic contained in mulBatchDesign input object.
}
\examples{
\dontrun{
## Using example data provided by MultiBaC package
data(multiyeast)
inputData <- inputData(A.rna, A.gro, B.rna, B.ribo, C.rna, C.par, batches = c(1,1,2,2,3,3),
omicNames = c("RNA", "GRO", "RNA", "RIBO", "RNA", "PAR"),
batchesNames = c("A", "B", "C"))

## Creating cond.factor
cond.factor = list("A" = c("Glu+", "Glu+", "Glu+", "Glu-", "Glu-", "Glu+"),
"B" = c("Glu+", "Glu+", "Glu-", "Glu-"),
"C" = c("Glu+", "Glu+", "Glu-", "Glu-"))

## Create PLS models
modelList <- genModelList(input, test.comp = 5)$modelList

## Generate missing omic data
missingOmics <- genMissingOmics(input, modelList, commonOmic = 1)

## Batch correction across omics

correctedOmics <- batchCorrection(missingOmics, cond.factor)$correctedOmics}
}
\references{
Nueda MJ, Ferrer A, Conesa A. ARSyN: A method for the identification and removal of systematic noise in multifactorial time course microarray experiments. Biostatistics. 2012;13(3):553-566. doi:10.1093/biostatistics/kxr042
}
