% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxfunctions.R
\name{batchEstimation}
\alias{batchEstimation}
\title{batchEstimation}
\usage{
batchEstimation(Xunfold, batch_factor)
}
\arguments{
\item{Xunfold}{Samples x genes matrix containing the common omic information for all batches (labs)}

\item{batch_factor}{Class factor indicating the batch of each sample in Xunfold}
}
\description{
batchEstimation
}
