\name{C.par}
\docType{data}
\alias{C.par}
\title{Protein binded mRNA data from yeast}
\description{
  This data set gives the protein binding information of 200 genes from yeast measured at two different
  conditions (glucose vs glucose starvation).
}
\usage{multiyeast}
\format{A matrix containing 4 rows and 200 variables.}
\source{Gene Expression Omnimus: GSE43747}
\references{
[1] Freeberg, M. A., Han, T., Moresco, J. J., Kong, A., Yang, Y. C., Lu, Z. J., Yates, J. R., & Kim, J. K. (2013). Pervasive and dynamic protein binding sites of the mRNA transcriptome in Saccharomyces cerevisiae. Genome biology, 14(2), R13. https://doi.org/10.1186/gb-2013-14-2-r13
}
\keyword{datasets}
