\name{C.rna}
\docType{data}
\alias{C.rna}
\title{Gene expression data from yeast}
\description{
  This data set gives the expression of 200 genes from yeast measured at two different
  conditions (glucose vs glucose starvation).
}
\usage{multiyeast}
\format{A matrix containing 4 rows and 200 variables.}
\source{Gene Expression Omnimus: GSE43747}
\references{
@article{PARCLIP,
	An = {23409723},
	Author = {Freeberg, Mallory A and Han, Ting and Moresco, James J and Kong, Andy and Yang, Yu-Cheng and Lu, Zhi John and Yates, John R and Kim, John K},
	Date = {2013/02/14},
	Date-Added = {2019-01-29 12:26:46 +0000},
	Date-Modified = {2019-01-29 12:26:46 +0000},
	Db = {PubMed},
	Doi = {10.1186/gb-2013-14-2-r13},
	Isbn = {1474-760X; 1465-6906},
	J1 = {Genome Biol},
	Journal = {Genome biology},
	L2 = {https://www.ncbi.nlm.nih.gov/pmc/PMC4053964/},
	Month = {02},
	Number = {2},
	Pages = {R13--R13},
	Publisher = {BioMed Central},
	Title = {Pervasive and dynamic protein binding sites of the mRNA transcriptome in Saccharomyces cerevisiae},
	Ty = {JOUR},
	U1 = {23409723{$[$}pmid{$]$}},
	U2 = {PMC4053964{$[$}pmcid{$]$}},
	U4 = {gb-2013-14-2-r13{$[$}PII{$]$}},
	Url = {https://www.ncbi.nlm.nih.gov/pubmed/23409723},
	Volume = {14},
	Year = {2013},
	Bdsk-Url-1 = {https://www.ncbi.nlm.nih.gov/pubmed/23409723},
	Bdsk-Url-2 = {http://dx.doi.org/10.1186/gb-2013-14-2-r13}}
}
\keyword{datasets}
