% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxfunctions.R, R/plsClass.R
\name{computeQ2}
\alias{computeQ2}
\title{computeQ2}
\usage{
computeQ2(modelObject, messages = TRUE)

computeQ2(modelObject, messages = TRUE)
}
\arguments{
\item{Y}{Response matrix with observed data}

\item{Yhat}{Response matrix with predicted data}

\item{...}{}
}
\description{
computeQ2

Title
}
