% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiBaC.R
\name{createPLSmodel}
\alias{createPLSmodel}
\title{createPLSmodel}
\usage{
createPLSmodel(X, Y, test.comp, it = 50, messages = TRUE,
  tol = 1e-08, scale = FALSE, center = TRUE, plot = TRUE)
}
\arguments{
\item{X}{matrix of predictor variables}

\item{Y}{matrix of response variables}

\item{test.comp}{maximum number of latent variables (or components) to test}

\item{it}{maximum number of iterations allowed for nipals algorithm}

\item{tol}{threshold value to validate nipals algorithm convergence}

\item{scale}{TRUE or FALSE. Whether X and Y matrices have to be scaled}

\item{center}{TRUE or FALSE. Whether X and Y matrices have to be centered}

\item{plot}{TRUE or FALSE. Whether to plot Q^2 across components.}
}
\value{
Instance of class "plsClass"
}
\description{
createPLSmodel
}
\examples{
\dontrun{
}
}
