% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiBaC.R
\name{genMissingOmics}
\alias{genMissingOmics}
\title{genMissingOmics}
\usage{
genMissingOmics(ListOfBatches, modelList, commonOmic)
}
\arguments{
\item{ListOfBatches}{A list. Object returned by inputData function}

\item{modelList}{A list of PLS models. Oject returned by genModelList function}

\item{commonOmic}{Name or index of the commonOmic between the batches. If the index is given, the common data matrix must be in the same possition in each batch.}
}
\value{
A type list object.
}
\description{
This function generates for all the batches the omic data they had not originally. This is the previous step to apply ARSyN correction.
}
\examples{
\dontrun{
## Using example data provided by MultiBaC package
data(multiyeast)
inputData <- inputData(A.rna, A.gro, B.rna, B.ribo, C.rna, C.par, batches = c(1,1,2,2,3,3),
omicNames = c("RNA", "GRO", "RNA", "RIBO", "RNA", "PAR"),
batchesNames = c("A", "B", "C"))

## Creating cond.factor
cond.factor = list("A" = c("Glu+", "Glu+", "Glu+", "Glu-", "Glu-", "Glu+"),
"B" = c("Glu+", "Glu+", "Glu-", "Glu-"),
"C" = c("Glu+", "Glu+", "Glu-", "Glu-"))

## Create PLS models
modelList <- genModelList(input, test.comp = 5)$modelList

## Generate missing omic data
missingOmics <- genMissingOmics(input, modelList, commonOmic = 1)

## Batch correction across omics

correctedOmics <- batchCorrection(missingOmics, cond.factor)$correctedOmics
}
}
