% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiBaC.R
\name{genModelList}
\alias{genModelList}
\title{genModelList}
\usage{
genModelList(ListOfBatches, test.comp, scale = FALSE, center = TRUE,
  crossval = NULL, commonOmic = 1)
}
\arguments{
\item{ListOfBatches}{A list. Object returned by inputData function.}

\item{test.comp}{Maximum number of components allowed in PLS models.}

\item{scale}{TRUE or FALSE. Whether X and Y matrices have to be scaled}

\item{center}{TRUE or FALSE. Whether X and Y matrices have to be centered}

\item{crossval}{Integer: number of cross-validation segments. The number of samples (rows of 'x') must be at least >= crossvalI.
If NULL (default) leave-one-out crossvalidation is performed}

\item{commonOmic}{Name or index of the commonOmic between the batches. If the index is given, the common data matrix must be in the same possition in each batch.}
}
\value{
A list with two slots:
\enumerate{
 \item{modelList: A list of PLS models.}
 \item{q2ofmodels: A list. Each slot contains a vector of Q^2 values of a PLS model.}
}
}
\description{
This function performs PLS models for every batch. A PLS model is generated for each non-common omic in each batch.
}
\examples{
\dontrun{
## Using example data provided by MultiBaC package
data(multiyeast)
inputData <- inputData(A.rna, A.gro, B.rna, B.ribo, C.rna, C.par, batches = c(1,1,2,2,3,3),
omicNames = c("RNA", "GRO", "RNA", "RIBO", "RNA", "PAR"),
batchesNames = c("A", "B", "C"))

## Creating cond.factor
cond.factor = list("A" = c("Glu+", "Glu+", "Glu+", "Glu-", "Glu-", "Glu+"),
"B" = c("Glu+", "Glu+", "Glu-", "Glu-"),
"C" = c("Glu+", "Glu+", "Glu-", "Glu-"))

## Create PLS models
modelList <- genModelList(input, test.comp = 5)$modelList

## Generate missing omic data
missingOmics <- genMissingOmics(input, modelList, commonOmic = 1)

## Batch correction across omics

correctedOmics <- batchCorrection(missingOmics, cond.factor)$correctedOmics
}
}
