% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxfunctions.R
\name{getData}
\alias{getData}
\title{getData}
\usage{
getData(ListOfBatches)
}
\arguments{
\item{ListOfBatches}{A list of MultiAssayExperiment elements. Object returned by inputData function.}
}
\value{
A list of matrices.
}
\description{
getData
}
\examples{
\dontrun{
## Using example data provided by MultiBaC package
data(multiyeast)
my_mbac <- createMbac (inputOmics = list(A.rna, A.gro, B.rna, B.ribo, C.rna, C.par),
                       batchFactor = c("A", "A", "B", "B", "C", "C"),
                       experimentalDesign = list("A" =  c("Glu+", "Glu+",
                       "Glu+", "Glu-", "Glu-", "Glu-"),
                       "B" = c("Glu+", "Glu+", "Glu-", "Glu-"),
                       "C" = c("Glu+", "Glu+", "Glu-", "Glu-")),
                       omicNames = c("RNA", "GRO", "RNA", "RIBO", "RNA", "PAR"))

omicData <- getData (my_mbac$ListOfBatches)}

}
