% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiBaC.R
\name{inputData}
\alias{inputData}
\title{inputData}
\usage{
inputData(..., batches, omicNames, batchesNames)
}
\arguments{
\item{...}{Matrices (features x samples) sepated by comma.}

\item{batches}{A vector or type factor. Indicates which batch each input matrix belongs to.}

\item{omicNames}{The names of input omics.}

\item{batchesNames}{The names of the different batches,}
}
\value{
A list containing one slot for each batch. Each slot contains a MultiAssayExperiment class object with the corresponding omics included.
}
\description{
This function creates a list object to be used by MultiBaC function from a set of matrix R objects.
}
\examples{
\dontrun{
#' ## Using example data provided by MultiBaC package
data(multiyeast)
inputData <- inputData(A.rna, A.gro, B.rna, B.ribo, C.rna, C.par, batches = c(1,1,2,2,3,3),
omicNames = c("RNA", "GRO", "RNA", "RIBO", "RNA", "PAR"),
batchesNames = c("A", "B", "C"))
}
}
