% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leverage.lims.R
\name{leverage.lims}
\alias{leverage.lims}
\title{Find leverage (importance of a variable in the PCA model) cut-off}
\usage{
leverage.lims(data = data, R = 100, FUN, Designa = Designa,
  Designb = Designb, Designc = NULL, Fac = c(1, 2, 2, 2), type = 2,
  alpha = 0.01, showvar = FALSE, showscree = FALSE)
}
\arguments{
\item{data}{Data matrix (p x n) that contains expression values of p genes (in rows) and n conditions (in columns).}

\item{R}{Number of repeats.}

\item{FUN}{Function to be called (ASCA.2f or ASCA.3f)}

\item{Designa}{First design matrix to pass to FUN.}

\item{Designb}{Second design matrix to pass to FUN.}

\item{Designc}{Third design matrix to pass to FUN.}

\item{Fac}{Vector with numbers of components to pass to FUN.}

\item{type}{Vector indicating whether the analyses of the model of the factors that interact with time (b.ab and c.ac) are studied jointly(1) or separately(2).}

\item{alpha}{Alpha value}

\item{showvar}{Show variability associated to each model or not.}

\item{showscree}{Show a screenplot.}
}
\value{
A list with elements:
\describe{
    \item{NullDistribution}{}
    \item{Cutoff}{}
    \item{Selection}{}
}
}
\description{
Find leverage (importance of a variable in the PCA model) cut-off
}
\examples{
\dontrun{
leverage.cutoff <- leverage.lims(data = data.example, R = 3, ASCA.3f, Designa = Designa,
    Designb = Designb, Designc = Designc,Fac = c(1,2,2,2,2,2,2,2), type = c(1,1,2),
    alpha = 0.95, showvar = FALSE, showscree = FALSE)
}
}
