% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxfunctions.R
\name{mbacClass}
\alias{mbacClass}
\title{Class MultiBaC}
\usage{
mbacClass(list, ...)
}
\arguments{
\item{list}{List of objects. At least 1 of them.
\enumerate{
    \item ListOfBatches: A list of MultiAssayExperiment objects (one per batch).
    \item commonOmic: Name of the common omic between the batches. It must be one of the names in omicNames argument. If NULL (default), the omic names that appears more times is selected as commonOmic.
    \item CorrectedData: Same structure than ListOfBatches but with the corrected data instead of the original.
    \item PLSmodels: PLS models created during MultiBaC method performance (one model per non-common omic data type).
    \item ARSyNmodels: ARSyN models created during MultiBaC performance (one per omic data type).
    \item InnerRelation: Table of class data.frame containing the inner correlation (i.e. correlation between the scores of X (t) and Y (u) matrices) for each PLS model across all components.
}}

\item{...}{Other attributes}
}
\value{
Instance of class "mbac"
}
\description{
Class MultiBaC
}
