% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiBaC.R
\docType{package}
\name{MultiBaC}
\alias{MultiBaC}
\alias{MultiBaC-package}
\title{MultiBaC}
\usage{
MultiBaC(X, YZ, test.comp, cond.factor, it = 50, tol = 1e-08,
  scale = FALSE, center = TRUE, showplot = TRUE, Fac = NULL,
  nm = FALSE)
}
\arguments{
\item{X}{list of common omics}

\item{YZ}{list of non-common omics}

\item{test.comp}{maximum number of components allowed}

\item{cond.factor}{list of experimental design conditions for each lab}

\item{it}{maximum number of iterations allowed for nipals algorithm}

\item{tol}{threshold value to validate nipals algorithm convergence}

\item{scale}{TRUE or FALSE. Whether X and Y matrices have to be scaled}

\item{center}{TRUE or FALSE. Whether X and Y matrices have to be centered}

\item{Fac}{Vector with numbers of components to extract in each submodel, by order:
\enumerate{
    \item Model a (time)
    \item Model b (second factor)
    \item Model c (third factor)
    \item Model ab (interaction) to not consider interations set to 0
    \item Model ac (interaction) to not consider interations set to 0
    \item Model bc (interaction) to not consider interations set to 0
    \item Model abc (interaction) to not consider interations set to 0
    \item Number components of residues
}}

\item{nm}{TRUE or FALSE. Whether to compute a final step to improve batch correction.}

\item{estBatchmagnitude}{TRUE or FALSE. Whether to show an estimation of batch effect magnitude or not}

\item{type}{Vector indicating whether the analyses of the model of the factors that interact with time (b.ab and c.ac) are studied jointly(1) or separately(2).}

\item{showvar}{Show variability associated to each model or not.}

\item{showscree}{Show a screenplot.}
}
\value{
A list with two slots
#' \describe{
    \item{X.star}{List of common omics matrices corrected}
    \item{YZ.star}{List of non-common omics matrices corrected}
}
}
\description{
Description
}
\examples{
Corrected_Matrices <- MultiBaC ( X = list("A" = rnaseq_a, "B" = rnaseq_b),
                                 YZ = list("A" = chipseq_a, "B" = metabolomics_b),
                                 test.comp = 3, cond.factor = design)
}
