% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxfunctions.R
\name{MultiBaC_plots}
\alias{MultiBaC_plots}
\title{MultiBaC_plot}
\usage{
MultiBaC_plots(q2values, ascamodels)
}
\arguments{
\item{q2values}{A list. Each slot contains a vector of Q^2 values of a PLS model.}

\item{ascamodels}{A list. Each slot contains a vector of cumulative explained variances of the batch estimation.}
}
\value{
Display two plots.
}
\description{
Display two plots summarizing MultiBaC steps.
}
