% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiBaC_class.R
\name{multibacClass}
\alias{multibacClass}
\title{Class MultiBaC}
\usage{
multibacClass(multibac, ...)
}
\value{
Instance of class "plsClass"
#' \describe{
    \item{X}{A list containing pls results of matrix X}
    \item{Y}{A list containing pls results of matrix Y}
    \item{Bhat}{Coefficients matrix of the PLS model}
    \item{scale}{TRUE or FALSE. Whether X and Y have been scaled or not}
    \item{center}{TRUE or FALSE. Whether X and Y have been centered or not}
}
}
\description{
Class MultiBaC
}
