% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxfunctions.R
\name{new_module}
\alias{new_module}
\title{new_module}
\usage{
new_module(YZ, missingOmics, cond.factor)
}
\arguments{
\item{YZ}{List of non-common omics}

\item{missingOmics}{List of predicted omics for each batch (lab)}

\item{cond.factor}{List of experimental condition factor type objects for each matrix in YZ}
}
\description{
new_module
}
\examples{

}
