% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxfunctions.R
\name{nipals}
\alias{nipals}
\title{nipals}
\usage{
nipals(X, a, it = 50, tol = 1e-10, center = FALSE, scale = FALSE)
}
\arguments{
\item{X}{Samples x genes numerical matrix.}

\item{a}{Number of latent variables (or components)}

\item{it}{Maximum number of iterations allowed to nipals algorithm}

\item{tol}{threshold value to validate nipals algorithm convergence}

\item{center}{TRUE or FALSE. Whether X and Y matrices have to be centered}

\item{scale}{TRUE or FALSE. Whether X and Y matrices have to be scaled}
}
\value{
A list containing two slots:
\describe{
    \item{T}{scores}
    \item{P}{loadings}
}
}
\description{
nipals
}
\examples{
pca <- nipals (X = t(assayData(my_counts)$exprs), a = 2, center = TRUE )
}
