% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsClass.R
\name{optimizeComponents}
\alias{optimizeComponents}
\title{optimizeComponents}
\usage{
optimizeComponents(modelObject)
}
\arguments{
\item{modelObject}{}
}
\description{
optimizeComponents
}
\examples{
\dontrun{
}
}
