% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA-GENES.R
\name{PCA.GENES}
\alias{PCA.GENES}
\title{Function to obtain principal components to a matrix that has more variables than individuals.}
\usage{
PCA.GENES(X)
}
\arguments{
\item{X}{Matrix that has on columns the genes considered as variables in the PCA analysis.}
}
\value{
A list with elements:
\describe{
    \item{eigen}{}
    \item{var.exp}{}
    \item{scores}{}
    \item{loadings}{}
}
}
\description{
Function to obtain principal components to a matrix that has more variables than individuals.
}
\examples{
\dontrun{
my.pca <- PCA.GENES(data.example)
}
}
