% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsClass.R
\name{plot.plsClass}
\alias{plot.plsClass}
\title{Create a method for function plot of plsClass}
\usage{
\method{plot}{plsClass}(modelObject, comp = c(1, 2),
  element = "scores", text = TRUE, ...)
}
\arguments{
\item{comp}{Vector of two dimension. Which components are plotted.}

\item{element}{Which element of PLS is plotted: "scores", "loadings", "weightings", "inner.scores" or "inner.weightings".}

\item{modellObject}{Instance of class "plsClass"}
}
\description{
Create a method for function plot of plsClass
}
