% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsClass.R
\name{pls}
\alias{pls}
\title{PLS model using nipals algorithm}
\usage{
pls(X, Y, a, it = 50, tol = 1e-08, scale = FALSE, center = TRUE)
}
\arguments{
\item{X}{matrix of predictor variables}

\item{Y}{matrix of response variables}

\item{a}{number of latent variables (or components)}

\item{it}{maximum number of iterations allowed for nipals algorithm}

\item{tol}{threshold value to validate nipals algorithm convergence}

\item{scale}{TRUE or FALSE. Whether X and Y matrices have to be scaled}

\item{center}{TRUE or FALSE. Whether X and Y matrices have to be centered}
}
\value{
Instance of class "plsClass"
}
\description{
PLS model using nipals algorithm
}
\examples{
\dontrun{
}
}
