% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsClass.R
\name{plsClass}
\alias{plsClass}
\title{Class PLS model}
\usage{
plsClass(pls_model, ...)
}
\value{
Instance of class "plsClass"
#' \describe{
    \item{X}{A list containing pls results of matrix X}
    \item{Y}{A list containing pls results of matrix Y}
    \item{Bhat}{Coefficients matrix of the PLS model}
    \item{scale}{TRUE or FALSE. Whether X and Y have been scaled or not}
    \item{center}{TRUE or FALSE. Whether X and Y have been centered or not}
}
}
\description{
Class PLS model
}
