% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsClass.R
\name{predict.plsClass}
\alias{predict.plsClass}
\title{Create a method for function predict of plsClass}
\usage{
\method{predict}{plsClass}(modelObject, newdata)
}
\arguments{
\item{modelObject}{Instance of class "plsClass"}

\item{newdata}{New observations to predict values of response matrix}
}
\description{
Create a method for function predict of plsClass
}
