% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sceeplot.R
\name{screeplot}
\alias{screeplot}
\title{Show a scree plot of an ASCA analysis}
\usage{
screeplot(asca)
}
\arguments{
\item{asca}{Result of ASCA analysis.}
}
\description{
Show a scree plot of an ASCA analysis
}
\examples{
\dontrun{
my.asca <- ASCA.3f(X = t(data.example), Designa = Designa,
    Designb = Designb, Designc = Designc, Fac= c(1, 2, 2, 2, 2, 2, 2, 2),
    type = c(1, 1, 2))

screeplot(my.asca)
}
}
