% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPE.lims.R
\name{SPE.lims}
\alias{SPE.lims}
\title{Find squared prediction error cut-off}
\usage{
SPE.lims(my.asca, alpha)
}
\arguments{
\item{my.asca}{Result of ASCA analysis.}

\item{alpha}{Alpha value.}
}
\description{
Find squared prediction error cut-off
}
\examples{
my.asca <- ASCA.3f(X = t(data.example), Designa = Designa,
    Designb = Designb, Designc = Designc, Fac= c(1, 2, 2, 2, 2, 2, 2, 2),
    type = c(1, 1, 2))

 SPE.cutoff <- SPE.lims(my.asca, alpha = 0.01)
}
