% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-function.R
\name{summary.mbac}
\alias{summary.mbac}
\title{summary.mbac}
\usage{
\method{summary}{mbac}(object, ...)
}
\arguments{
\item{object}{An object of class mbac.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
Custom mbac object structure.
}
\description{
Displays the structure and the content of the object of class mbac.
}
\examples{
data('multiyeast')

my_mbac <- createMbac (inputOmics = list(A.rna, A.gro, B.rna, B.ribo, C.rna, C.par),
                       batchFactor = c("A", "A", "B", "B", "C", "C"),
                       experimentalDesign = list("A" =  c("Glu+", "Glu+",
                       "Glu+", "Glu-", "Glu-", "Glu-"),
                       "B" = c("Glu+", "Glu+", "Glu-", "Glu-"),
                       "C" = c("Glu+", "Glu+", "Glu-", "Glu-")),
                       omicNames = c("RNA", "GRO", "RNA", "RIBO", "RNA", "PAR"),
                       commonOmic = "RNA")
summary(my_mbac)

}
