% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCA2f.R
\name{ASCA.2f}
\alias{ASCA.2f}
\title{#' ASCA for 2 experimental factors}
\usage{
ASCA.2f(X = X, Designa = Designa, Designb = Designb,
  Designc = NULL, Fac = c(1, 2, 2, 2), type = 1, showvar = TRUE,
  showscree = TRUE)
}
\arguments{
\item{X}{Data matrix (p x n) that contains expression values of n genes (in columns) and p conditions (in rows).}

\item{Designa}{Matrix (p x I) that contains 0's and 1's for the time-points in the experiment.}

\item{Designb}{Matrix (p x J) experimental group.}

\item{Designc}{Unused argument.}

\item{Fac}{Vector with numbers of components to extract in each submodel, by order:
\enumerate{
    \item Model a (time)
    \item Model b (second factor)
    \item Model ab (interaction) to not consider interations set to 0
    \item Number components of residues
}}

\item{type}{Number (1 or 2) to indicate if the analyses of the model b.ab is studied jointly(1) or separately(2).}

\item{showvar}{Show variability associated to each model or not.}

\item{showscree}{Show a screenplot.}
}
\value{
A list containing as many slots as computed models, each one with the following values:
\describe{
   \item{data}{data matrix used for PCA on that factor, corresponds to the estimated effects for each level of the factor}
   \item{scores}{corresponding to the PCA of that factor}
   \item{loadings}{corresponding to the PCA of that factor}
   \item{var.exp}{explained variability for each component in the PCA o that factor}
   \item{X}{the data matrix associated to each model}
   \item{TP}{the reconstructed data matrix corresponding that factor}
   \item{E}{residuals for the PCA model of that factor}
   \item{leverage}{gene leverages for the computed PCA model}
   \item{SPE}{gene SPEs for the computed PCA model}
 }
}
\description{
#' ASCA for 2 experimental factors
}
\examples{

my.asca <- ASCA.2f(X = t(data.example), Designa = Designa,
    Designb = Designb, Designc = NULL, Fac = c(1,2,2,2), type= 1)
}
