\name{B.rna}
\docType{data}
\alias{B.rna}
\title{Gene expression data from yeast}
\description{
  This data set gives the expression of 200 genes from yeast measured at two different
  conditions (glucose vs glucose starvation).
}
\usage{multiyeast}
\format{A matrix containing 4 rows and 200 variables.}
\source{Gene Expression Omnimus: GSE56622}
\references{
[1] Zid, B. M., & O'Shea, E. K. (2014). Promoter sequences direct cytoplasmic localization and translation of mRNAs during starvation in yeast. Nature, 514(7520), 117–121. https://doi.org/10.1038/nature13578
}
\keyword{datasets}

