% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiBaC.R
\name{createPLSmodel}
\alias{createPLSmodel}
\title{createPLSmodel}
\usage{
createPLSmodel(omicslist, test.comp, messages = TRUE, scale = FALSE,
  center = TRUE, crossval, regressor)
}
\arguments{
\item{omicslist}{A list of length 2. Each slot must contain a data matrix (features x samples).}

\item{test.comp}{maximum number of latent variables (or components) to test}

\item{messages}{If TRUE, messages about the algorithm steps will be displayed in console window.}

\item{scale}{TRUE or FALSE. Whether X and Y matrices have to be scaled}

\item{center}{TRUE or FALSE. Whether X and Y matrices have to be centered}

\item{crossval}{Integer: number of cross-validation segments. The number of samples (rows of 'x') must be at least >= crossvalI.
If NULL (default) leave-one-out crossvalidation is performed}

\item{regressor}{Integer (1 or 2): Which of the matrices is the X matrix for PLS model.}
}
\value{
Instance of class "opls". See package ropls. The final model has the optimal number of components used for prediction.
}
\description{
This function creates a PLS model between two omics data matrices. It also performs an optimization of the
number of component of the model that maximize the Q^2 as a measure of prediction performance.
}
\examples{
\dontrun{
## Using example data provided by MultiBaC package
data(example)

omicsList <- list("RNA" = A.rna, "GRO" = A.gro)

## Create model
plsModel <- createPLSmodel(omicsList, crossval = NULL, test.comp = 3,
scale = FALSE, center = TRUE, regressor = 1)
}
}
