% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxfunctions.R
\name{createPLSmodel}
\alias{createPLSmodel}
\title{createPLSmodel}
\usage{
createPLSmodel(omicslist, test.comp, messages = TRUE, scale = FALSE,
  center = TRUE, crossval, regressor, showinfo = TRUE)
}
\arguments{
\item{omicslist}{A list of length 2. Each slot must contain a data matrix (features x samples).}

\item{test.comp}{Maximum number of components allowed in PLS models. If NULL (default), the minimal effective rank of the matrices is used as the maximum number of components.}

\item{messages}{If TRUE, messages about the algorithm steps will be displayed in console window.}

\item{scale}{Logical. Whether X and Y matrices must be scaled. By default, FALSE.}

\item{center}{Logical. Whether X and Y matrices must be centered. By default, TRUE.}

\item{crossval}{Integer indicating the number of cross-validation segments. The number of samples (rows of 'x') must be at least >= crossvalI. If NULL (default), a leave-one-out crossvalidation is conducted.}

\item{regressor}{Integer (1 or 2): Which of the matrices is the X matrix for PLS model.}

\item{showinfo}{Logical. Whether to show function process in prompt.}
}
\value{
Instance of class "opls". See package ropls. The final model has the optimal number of components used for prediction.
}
\description{
This function creates a PLS model between two omics data matrices. It also performs an optimization of the
number of component of the model that maximize the Q^2 as a measure of prediction performance.
}
\examples{
\dontrun{
## Using example data provided by MultiBaC package
data(example)

omicsList <- list("RNA" = A.rna, "GRO" = A.gro)

## Create model
plsModel <- createPLSmodel(omicsList, crossval = NULL, test.comp = 3,
scale = FALSE, center = TRUE, regressor = 1)
}
}
