% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxfunctions.R
\name{getData}
\alias{getData}
\title{getData}
\usage{
getData(ListOfBatches)
}
\arguments{
\item{ListOfBatches}{A list of MultiAssayExperiment elements. Object returned by inputData function.}
}
\value{
A list of lists.
}
\description{
getData
}
\examples{
\dontrun{
## Using example data provided by MultiBaC package
data(multiyeast)
inputData <- readData(A.rna, A.gro, B.rna, B.ribo, C.rna, C.par, batches = c(1,1,2,2,3,3),
omicNames = c("RNA", "GRO", "RNA", "RIBO", "RNA", "PAR"),
batchesNames = c("A", "B", "C"))

omicData <- getData (inputData)}
}
