% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiBaC.R
\docType{package}
\name{MultiBaC}
\alias{MultiBaC}
\alias{MultiBaC-package}
\title{MultiBaC}
\usage{
MultiBaC(ListOfBatches, test.comp, cond.factor, commonOmic = 1,
  scale = FALSE, center = TRUE, showplot = TRUE, crossval = NULL,
  Fac = NULL)
}
\arguments{
\item{ListOfBatches}{A list. Object returned by inputData function.}

\item{test.comp}{maximum number of components allowed}

\item{cond.factor}{A list with one slot for each batch containing the common experimental setting.}

\item{commonOmic}{Name or index of the commonOmic between the batches. If the index is given, the common data matrix must be in the same possition in each batch.}

\item{scale}{TRUE or FALSE. Whether X and Y matrices have to be scaled}

\item{center}{TRUE or FALSE. Whether X and Y matrices have to be centered}

\item{showplot}{If TRUE, the Q^2 and the Fac optimization will be plotted.}

\item{crossval}{Integer: number of cross-validation segments. The number of samples (rows of 'x') must be at least >= crossvalI.
If NULL (default) leave-one-out crossvalidation is performed}

\item{Fac}{Vector with numbers of components to extract in each submodel, by order:
\enumerate{
    \item Model a (time)
    \item Model b (second factor)
    \item Model c (third factor)
    \item Model ab (interaction) to not consider interations set to 0
    \item Model ac (interaction) to not consider interations set to 0
    \item Model bc (interaction) to not consider interations set to 0
    \item Model abc (interaction) to not consider interations set to 0
    \item Number components of residues
}
If NULL (default) the function optimize each value.}
}
\value{
A list with as many slots as the given number of batches.
Returns the same structure than the input data "ListOfBatches" but with batch-corrected data.
}
\description{
Description

MultiBaC performs a multi-omic, multi-batch correction
}
\examples{
\dontrun{
## Using example data provided by MultiBaC package
data(example)
inputData <- readData(A.rna, A.gro, B.rna, B.ribo, C.rna, C.par, batches = c(1,1,2,2,3,3),
omicNames = c("RNA", "GRO", "RNA", "RIBO", "RNA", "PAR"),
batchesNames = c("A", "B", "C"))

## Creating cond.factor
cond.factor = list("A" = c("Glu+", "Glu+", "Glu+", "Glu-", "Glu-", "Glu+"),
"B" = c("Glu+", "Glu+", "Glu-", "Glu-"),
"C" = c("Glu+", "Glu+", "Glu-", "Glu-"))

## Applying MultiBaC
Corrected_Matrices <- MultiBaC (inputData, test.comp = 5,
cond.factor = cond.factor, commonOmic = "RNA")}

}
